OswAppWaterLevel: rename water-level
